pyproject.toml
[project]
name = "hyphae"
version = "2.0.0"
description = "Privacy-first scientific research copilot with hybrid local/cloud AI"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Team Darphie"},
]
keywords = ["research", "ai", "privacy", "local-first", "gemini", "cactus"]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.9",
    "pymupdf>=1.27",
    "google-genai>=1.64",
    "numpy>=1.25",
    "faiss-cpu>=1.7",
    "pydantic>=2.0",
    "httpx>=0.27",
    "bcrypt>=4.0",
    "sentence-transformers>=2.2",
    "trafilatura>=1.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0",
    "pytest-cov>=5.0",
    "requests>=2.32",
    "ruff>=0.6",
    "mypy>=1.10",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "web", "."]
addopts = "-v --tb=short"

[tool.setuptools.packages.find]
where = ["src", "web"]

# ── Ruff: linter + formatter ────────────────────────────────────────────────
[tool.ruff]
line-length = 110
target-version = "py312"
extend-exclude = ["cactus", ".venv", "build", "dist"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "RUF", # ruff-specific
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # function calls in argument defaults (FastAPI Depends/Query)
    "B904",  # raise from inside except (verbose for HTTPException)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["B011", "E402"]
"scripts/**/*.py" = ["E402"]
"examples/**/*.py" = ["E402"]
"main.py" = ["E402"]
"web/app.py" = ["E402"]  # bootstrap() must run before imports
"web/bootstrap.py" = ["E402"]

# ── Mypy: type checking ─────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.12"
files = ["src", "web"]
exclude = ["cactus", ".venv", "build", "web/notebook/__pycache__"]
ignore_missing_imports = true
follow_imports = "silent"
warn_unused_ignores = false
warn_redundant_casts = true
check_untyped_defs = false  # gradual typing — tighten over time
no_implicit_optional = true

[[tool.mypy.overrides]]
module = ["cactus", "cactus.*", "faiss", "faiss.*", "trafilatura", "fitz",
          "sentence_transformers", "google.*"]
ignore_missing_imports = true

# ── Coverage ────────────────────────────────────────────────────────────────
[tool.coverage.run]
branch = true
source = ["src", "web"]
omit = ["*/__pycache__/*", "tests/*", "cactus/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
